feat(AvatarLabel): clean Slots
